+++ src/main/resources/templates/Create.html
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<!--
+Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+Click nbfs://nbhost/SystemFileSystem/Templates/Other/html.html to edit this template
+-->
+<html>
+    <head>
+        <title>TODO supply a title</title>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+         <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
+    </head>
+    <body>
+         <form action="#" th:action="@{/create}" th:object="${udto}" method="post">
+             <h3 class="row justify-content-center"> Create User</h3>
+            <div class="row justify-content-center">
+                <div class="col-3">
+                    <p>Name: <input class="form-control" type="text" th:field="*{username}" /></p>
+                </div>
+                <div class="col-3">
+                    <p>Password: <input class="form-control" type="text" th:field="*{password}" /></p>
+                </div>
+            </div>
+            <div class="row justify-content-center">
+                <div class="col-2">
+                    <p>Create Admin: <input class="form-check-input" type="checkbox" th:field="*{admin}" /></p>  
+                </div>
+            </div>
+            <div class="row justify-content-center">
+                  <div class="col-2">
+            <p><input class="btn btn-primary" type="submit" value="Create User" /> <input class="btn btn-warning" type="reset" value="Reset" /></p>
+              </div >
+            </div>
+        </form>
+        
+         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
+        <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js" integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3" crossorigin="anonymous"></script>
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.min.js" integrity="sha384-mQ93GR66B00ZXjt0YO5KlohRA5SY2XofN4zfuZxLkoj1gXtW8ANNCe9d5Y3eG5eD" crossorigin="anonymous"></script>
+    </body>
+</html>
