+++ src/main/resources/templates/Login.html
@@ -24,45 +24,12 @@ Click nbfs://nbhost/SystemFileSystem/Templates/Other/html.html to edit this temp
                 </div>
             </div>
             <div class="row justify-content-center">
-                <div class="col-2">
-                    <p>Create New: <input class="form-check-input" type="checkbox" th:field="*{create}" /></p>
-                </div>
-                <div class="col-2">
-                    <p>Create Admin: <input class="form-check-input" type="checkbox" th:field="*{admin}" /></p>  
-                </div>
-            </div>
-            <div class="row justify-content-center">
                   <div class="col-2">
             <p><input class="btn btn-primary" type="submit" value="Login" /> <input class="btn btn-warning" type="reset" value="Reset" /></p>
               </div >
             </div>
         </form>
-        
-         <form action="#" th:action="@{/login}" th:object="${udto}" method="post">
-             <h3 class="row justify-content-center"> Create User</h3>
-            <div class="row justify-content-center">
-                <div class="col-3">
-                    <p>Name: <input class="form-control" type="text" th:field="*{username}" /></p>
-                </div>
-                <div class="col-3">
-                    <p>Password: <input class="form-control" type="text" th:field="*{password}" /></p>
-                </div>
-            </div>
-            <div class="row justify-content-center">
-                <div class="col-2">
-                    <p>Create New: <input class="form-check-input" type="checkbox" th:field="*{create}" /></p>
-                </div>
-                <div class="col-2">
-                    <p>Create Admin: <input class="form-check-input" type="checkbox" th:field="*{admin}" /></p>  
-                </div>
-            </div>
-            <div class="row justify-content-center">
-                  <div class="col-2">
-            <p><input class="btn btn-primary" type="submit" value="Login" /> <input class="btn btn-warning" type="reset" value="Reset" /></p>
-              </div >
-            </div>
-        </form>
-        
+       
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
         <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js" integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3" crossorigin="anonymous"></script>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.min.js" integrity="sha384-mQ93GR66B00ZXjt0YO5KlohRA5SY2XofN4zfuZxLkoj1gXtW8ANNCe9d5Y3eG5eD" crossorigin="anonymous"></script>
